Add prodcut ID validation

+++ Backend/controllers/product.controller.js
@@ -30,14 +30,18 @@ export const createProduct = async (req, res) => {
 
 export const updateProduct = async (req, res) => {
   const { id } = req.params; 
-  const updatedProduct = req.body; // get the updated product data from the request body
+  const updatedData = req.body; // get the updated product data from the request body
+
+  if (!mongoose.Types.ObjectId.isValid(id)) { //checking if the id is valid
+  return res.status(400).json({ success: false, message: 'Invalid product ID' }); 
+}
 
   try {
-    const product = await Product.findByIdAndUpdate(id, updatedProduct, { new: true }); //findByIdAndUpdate returns the default document before update
-    if (!product) {
-      return res.status(404).json({ success: false, message: 'Product not found' });
+    const updatedProduct = await Product.findByIdAndUpdate(id, updatedData, { new: true }); //findByIdAndUpdate returns the default document before update
+    if (!updatedProduct) {
+      return res.status(404).json({ success: false, message: 'Product not found' }); //In case the ID is valid but not in DB
     }
-    res.status(200).json({ success: true, data: product });
+    res.status(200).json({ success: true, data: updatedProduct });
   } 
   catch (error) {
     console.log("Error updating product:", error);
@@ -49,9 +53,18 @@ export const deleteProduct = async (req, res) => {
 
   const {id}= req.params; // get the product id from the request parameters
   console.log("Product ID to delete:", id);
+  
+  if (!mongoose.Types.ObjectId.isValid(id)) { //checking if the id is valid
+    return res.status(400).json({ success: false, message: 'Invalid product ID' }); 
+  }
 
   try {
-    await Product.findByIdAndDelete(id); // delete the product from the database
+    const deletedProduct = await Product.findByIdAndDelete(id); // delete the product from the database
+    
+    if (!deletedProduct) {
+      return res.status(404).json({ success: false, message: 'Product not found' });
+    }
+    
     res.status(200).json({ success: true, message: 'Product deleted successfully' });
   } 
   catch (error) {
